refactor: 기존 pointManage() 메서드가 포인트 적립내역, 포인트 사용내역 모두 조회해서 보여주므로 2개의 메서드로 분리

+++ src/main/java/com/java/guesthouse/guestdelluna/domain/DellunaDao.java
@@ -354,7 +354,7 @@ public class DellunaDao {
         return sqlSessionTemplate.selectOne("dao.dellunaMapper.getCountUse", memberCode);
     }
 
-    public List<PointAccumulateDto> myAccuPoint(int memberCode, int startRow, int endRow) {
+    public List<PointAccumulateDto> myAccuPoint(Long memberCode, int startRow, int endRow) {
 
         Map<String, Object> map = new HashMap<>();
         map.put("memberCode", memberCode);
@@ -364,7 +364,7 @@ public class DellunaDao {
         return sqlSessionTemplate.selectList("dao.dellunaMapper.myAccuPoint", map);
     }
 
-    public List<PointUse> myUsePoint(int memberCode, int startRow, int endRow) {
+    public List<PointUse> myUsePoint(Long memberCode, int startRow, int endRow) {
 
         Map<String, Object> map = new HashMap<>();
         map.put("memberCode", memberCode);
